requirements.txt
Django==5.2.5
djangorestframework
requests==2.32.3
beautifulsoup4==4.13.3
django-cors-headers
djangorestframework-simplejwt
gunicorn
psycopg[binary]

#docker compose -f docker-compose.prod.yml build --no-cache --progress=plain